+++ libs/src/assets/icons/icons.js
@@ -1,0 +1,111 @@
+import SvgAccount from './SvgAccount.svg';
+import SvgAdd from './SvgAdd.svg';
+import SvgAir from './SvgAir.svg';
+import SvgArrowDown from './SvgArrowDown.svg';
+import SvgArrowDropDown from './SvgArrowDown.svg';
+import SvgArrowLeft from './SvgArrowLeft.svg';
+import SvgBarChart from './SvgBarChart.svg';
+import SvgBusy from './SvgBusy.svg';
+import SvgCalendar from './SvgCalendar.svg';
+import SvgCheck from './SvgCheck.svg';
+import SvgClock from './SvgClock.svg';
+import SvgClose from './SvgClose.svg';
+import SvgCurtain from './SvgCurtain.svg';
+import SvgDatabase from './SvgDatabase.svg';
+import SvgDisableCircle from './SvgDisableCircle.svg';
+import SvgDoor from './SvgDoor.svg';
+import SvgErrorCircle from './SvgErrorCircle.svg';
+import SvgFacebook from './SvgFacebook.svg';
+import SvgFan from './SvgFan.svg';
+import SvgFavorite from './SvgFavorite.svg';
+import SvgFirstLastPage from './SvgFirstLastPage.svg';
+import SvgGoogle from './SvgGoogle.svg';
+import SvgHome from './SvgHome.svg';
+import SvgHumidity from './SvgHumidity.svg';
+import SvgIdle from './SvgIdle.svg';
+import SvgInfo from './SvgInfo.svg';
+import SvgInfoCircle from './SvgInfoCircle.svg';
+import SvgLanguage from './SvgLanguage.svg';
+import SvgLight from './SvgLight.svg';
+import SvgLine from './SvgLine.svg';
+import SvgLocation from './SvgLocation.svg';
+import SvgLock from './SvgLock.svg';
+import SvgLogout from './SvgLogout.svg';
+import SvgMail from './SvgMail.svg';
+import SvgMinus from './SvgMinus.svg';
+import SvgMore from './SvgMore.svg';
+import SvgMoreVert from './SvgMoreVert.svg';
+import SvgNavigateArrow from './SvgNavigateArrow.svg';
+import SvgNotification from './SvgNotification.svg';
+import SvgOffline from './SvgOffline.svg';
+import SvgOnline from './SvgOnline.svg';
+import SvgPlus from './SvgPlus.svg';
+import SvgPower from './SvgPower.svg';
+import SvgQuestionCircle from './SvgQuestionCircle.svg';
+import SvgSearch from './SvgSearch.svg';
+import SvgSettings from './SvgSettings.svg';
+import SvgSuccessCircle from './SvgSuccessCircle.svg';
+import SvgTemperature from './SvgTemperature.svg';
+import SvgUser from './SvgUser.svg';
+import SvgVisibility from './SvgVisibility.svg';
+import SvgVisibilityOff from './SvgVisibilityOff.svg';
+import SvgWarning from './SvgWarning.svg';
+import SvgWarningTri from './SvgWarningTri.svg';
+import SvgWindow from './SvgWindow.svg';
+
+export default {
+    SvgAccount,
+    SvgAdd,
+    SvgAir,
+    SvgArrowDown,
+    SvgArrowDropDown,
+    SvgArrowLeft,
+    SvgBarChart,
+    SvgBusy,
+    SvgCalendar,
+    SvgCheck,
+    SvgClock,
+    SvgClose,
+    SvgCurtain,
+    SvgDatabase,
+    SvgDisableCircle,
+    SvgDoor,
+    SvgErrorCircle,
+    SvgFacebook,
+    SvgFan,
+    SvgFavorite,
+    SvgFirstLastPage,
+    SvgGoogle,
+    SvgHome,
+    SvgHumidity,
+    SvgIdle,
+    SvgInfo,
+    SvgInfoCircle,
+    SvgLanguage,
+    SvgLight,
+    SvgLine,
+    SvgLocation,
+    SvgLock,
+    SvgLogout,
+    SvgMail,
+    SvgMinus,
+    SvgMore,
+    SvgMoreVert,
+    SvgNavigateArrow,
+    SvgNotification,
+    SvgOffline,
+    SvgOnline,
+    SvgPlus,
+    SvgPower,
+    SvgQuestionCircle,
+    SvgSearch,
+    SvgSettings,
+    SvgSuccessCircle,
+    SvgTemperature,
+    SvgUser,
+    SvgVisibility,
+    SvgVisibilityOff,
+    SvgWarning,
+    SvgWarningTri,
+    SvgWindow
+};